requirements.txt
requests>=2.20.0
Flask==0.12.3
pymongo==3.4.0
flask_pymongo==0.5.1
Flask_JWT_Extended==3.7.2
Flask_Bcrypt==0.7.1
jsonschema==2.5.1
pyjwt








